requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
#backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
cssselect2==0.6.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.12
django-allauth==0.49.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-environ==0.8.1
django-filter==21.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-modeltranslation==0.17.5
django-mptt==0.13.4
django-notifications-hq==1.6.0
django-phonenumber-field==6.1.0
django-qr-code==3.0.0
django-random-id-model==0.1.1
django-resized==0.3.11
django-rosetta==0.9.8
django-soft-delete==1.0.15
django-sort-order-field==1.2
django-storages==1.12.3
django-taggit==2.1.0
django-timezone-field==4.2.3
django-tinymce==3.4.0
django-tz-detect==0.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
exponent-server-sdk==2.0.0
fonttools==4.32.0
future==0.18.2
html5lib==1.1
idna==3.3
importlib-metadata==4.11.1
jsonfield==3.1.0
jsonpickle==3.0.1
kombu==5.2.3
lxml==4.8.0
Markdown==3.3.6
mixins==0.1.4
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
pdfkit==1.0.0
phonenumbers==8.12.43
Pillow==9.0.1
polib==1.1.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.14.1
pydyf==0.1.2
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==3.0.7
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyuca==1.2
PyYAML==6.0
qrcode==7.3.1
random-username==1.0.2
redis==4.1.4
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
rollbar==0.16.2
segno==1.4.1
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
swapper==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
zipp==3.7.0
zopfli==0.2.1
